+++ bika/lims/browser/widgets/analysisprofileanalyseswidget.py
@@ -188,8 +188,6 @@ class AnalysisProfileAnalysesView(BikaListingView):
         return columns
 
     def folderitems(self):
-        """TODO: Refactor to non-classic mode
-        """
         items = super(AnalysisProfileAnalysesView, self).folderitems()
         self.categories.sort()
         return items