+++ bika/lims/browser/widgets/artemplateanalyseswidget.py
@@ -197,8 +197,6 @@ class ARTemplateAnalysesView(BikaListingView):
         return columns
 
     def folderitems(self):
-        """TODO: Refactor to non-classic mode
-        """
         items = super(ARTemplateAnalysesView, self).folderitems()
         self.categories.sort()
         return items