+++ bika/lims/catalog/catalog_basic_template.py
@@ -55,5 +55,4 @@ BASE_CATALOG_COLUMNS = [
     "allowedRolesAndUsers",
     "created",
     "Creator",
-    "getObjectWorkflowStates",
 ]