+++ bika/lims/controlpanel/bika_labcontacts.py
@@ -22,6 +22,7 @@ import collections
 
 from Products.ATContentTypes.content import schemata
 from Products.Archetypes import atapi
+from bika.lims import api
 from bika.lims import bikaMessageFactory as _
 from bika.lims.browser.bika_listing import BikaListingView
 from bika.lims.config import PROJECTNAME
@@ -125,7 +126,7 @@ class LabContactsView(BikaListingView):
             the template
         :index: current index of the item
         """
-
+        obj = api.get_object(obj)
         fullname = obj.getFullname()
         if fullname:
             item["Fullname"] = fullname